Add initial IRCv3 support
This adds some very initial IRCv3 support and SAS authentication.

The CAP negotiation is initialised before login is attempted, and this negotiation is controlled by a State machine that tracks what the current negotiation state and transitions to the next state once completed.

The main command processing switch now reacts to CAP and the 9xx SASL commands and where appropriate delegates the processing into the State machine.

Also added is some the UseSasl option into the connection options, and the connection options are also now exposed as a public property on the client, so that they can be access from within the state machine.